+++ src/backend/joanie/settings.py
@@ -159,6 +159,7 @@ class Base(Configuration):
         "django.middleware.csrf.CsrfViewMiddleware",
         "django.contrib.auth.middleware.AuthenticationMiddleware",
         "django.contrib.messages.middleware.MessageMiddleware",
+        "django.contrib.sites.middleware.CurrentSiteMiddleware",
         "dockerflow.django.middleware.DockerflowMiddleware",
     ]